requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django
django-extensions==2.2.5
django-rest-framework
djangorestframework
#psycopg2==2.8.4
#pygraphviz==1.5
python-decouple==3.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
